FILE: moves mock-data into mappers folder
