+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/mappers/map-ingredient-type-dto-to-ingredient-type.spec.ts
@@ -1,5 +1,5 @@
-import { INGREDIENT_DTOS, INGREDIENT_TYPE_DTOS } from '../dtos/mock-data';
 import { mapIngredientTypeDtoToIngredientType } from './map-ingredient-type-dto-to-ingredient-type';
+import { INGREDIENT_DTOS, INGREDIENT_TYPE_DTOS } from './mock-data';
 
 describe('mapIngredientTypeDtoToIngredientType', () => {
   it('returns an ingredientType when passed ingredients', () => {
