+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/mappers/map-meal-dto-to-meal.spec.ts
@@ -1,12 +1,12 @@
+import { mapDishDtoToDish } from './map-dish-dto-to-dish';
+import { mapMealDtoToMeal } from './map-meal-dto-to-meal';
 import {
   INGREDIENT_DTOS,
   MEAL_DTO,
   PIZZA_DTO,
   SALAD_DTO,
   TAG_DTOS,
-} from '../dtos/mock-data';
-import { mapDishDtoToDish } from './map-dish-dto-to-dish';
-import { mapMealDtoToMeal } from './map-meal-dto-to-meal';
+} from './mock-data';
 
 describe('mapMealDtoToMeal', () => {
   it('returns a meal when passed a mealDto, dishes, and tags', () => {
