+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/mappers/map-menu-dto-to-menu.spec.ts
@@ -1,12 +1,12 @@
+import { mapDishDtoToDish } from './map-dish-dto-to-dish';
+import { mapMenuDtoToMenu } from './map-menu-dto-to-menu';
 import {
   MENU_DTO,
   PIZZA_DTO,
   INGREDIENT_DTOS,
   TAG_DTOS,
   SALAD_DTO,
-} from '../dtos/mock-data';
-import { mapDishDtoToDish } from './map-dish-dto-to-dish';
-import { mapMenuDtoToMenu } from './map-menu-dto-to-menu';
+} from './mock-data';
 
 describe('mapMenuDtoToMenu', () => {
   it('returns a menu when passed a menuDto, dishes, and userPreferences', () => {
